Feed url format fix + WebHelper deprecation
