+++ AndroidApp/app/src/main/res/values/strings.xml
@@ -107,7 +107,6 @@
     <string name="error_url">Error reading from url.</string>
     <string name="error_host">No internet connection or bad host url.</string>
 
-    <string name="rsslocale_news">News</string>
 
     <string name="sendmail">Send email</string>
     <string name="noemailfound">No email app was found. Send feed back to niilo.poutanen@gmail.com</string>
